Changes to hierarchy and changes to some classes
I have changed the hierarchy of the project to have 3 source files: 1.) mian - it has the main function with the data types it needs to run properly. 2.) models - which contains all the model classes that model our game object. 3.) view contains the classes that will take care of any consoleIO, read and write to files.

Other changes have been made to the program to improve functionality. A new main class has been added and just a skeleton to start