Fixed Print and Prompt. Remove Unnecessary classes.
This commit changes some methods in Print and Prompt to static so we can use them anywhere in the game without creating an instance. Also, I removed ViewLogic and readWriter.  Viewlogic was Just confusing now we have two classes responsible for view  Print and Prompt. I do not think we will ever read or write files from the program so  I removed it.