Add logic and Comment Game/ Update functionality in Print
In this commit, I made sure that the present statements are updated and clearly print to a user the state of the game. Also now PrintMeldtoUser () checks if the meld is empty inside it to print a message accordingly. I have added the ability of a player to draw a tile in the game if he does not play any melds on that turn. Also right after playing if a player has 0 tiles the players is considered the winner and the game ends