Add Prompt for each turn, Add ability to turn of printing for computers
In this commit I added the ability to prompt for each turn so that we can see what each player plays. This can be turned on and off by a boolean, Also, I added the ability to be able to turn on and off printing the rack and melds for all computers. This is controlled by a boolean in game also.