Auto stash before merge of "dev" and "origin/dev"
Many changes to prepare for strategy 4 to work. Changed a method in table to send an int[] parameters to tableInfo of the players hand count instead of just an int parameter of the lowest count. Updated strategy 3 to take of the count and find the lowest one. Now strategy 4 will use this array to find the total count of racks for each player.